Move headers to correct location
