+++ include/dlaf/auxiliary/norm.h
@@ -15,7 +15,7 @@
 #include "dlaf/communication/communicator_grid.h"
 #include "dlaf/lapack/enum_output.h"
 #include "dlaf/lapack_tile.h"
-#include "dlaf/matrix.h"
+#include "dlaf/matrix/matrix.h"
 #include "dlaf/types.h"
 #include "dlaf/util_matrix.h"
