+++ include/dlaf/auxiliary/norm/mc.h
@@ -18,7 +18,7 @@
 #include "dlaf/communication/communicator_grid.h"
 #include "dlaf/communication/sync/reduce.h"
 #include "dlaf/lapack_tile.h"
-#include "dlaf/matrix.h"
+#include "dlaf/matrix/matrix.h"
 #include "dlaf/matrix/distribution.h"
 #include "dlaf/types.h"
 #include "dlaf/util_matrix.h"
