+++ include/dlaf/solver/triangular.h
@@ -11,7 +11,7 @@
 
 #include <blas.hh>
 #include "dlaf/communication/communicator_grid.h"
-#include "dlaf/matrix.h"
+#include "dlaf/matrix/matrix.h"
 #include "dlaf/solver/triangular/mc.h"
 #include "dlaf/types.h"
 #include "dlaf/util_matrix.h"
