+++ include/dlaf/solver/triangular/mc.h
@@ -22,7 +22,7 @@
 #include "dlaf/communication/executor.h"
 #include "dlaf/communication/functions_sync.h"
 #include "dlaf/lapack_tile.h"
-#include "dlaf/matrix.h"
+#include "dlaf/matrix/matrix.h"
 #include "dlaf/matrix/distribution.h"
 #include "dlaf/solver/triangular/api.h"
 #include "dlaf/util_matrix.h"
