+++ include/dlaf/util_matrix.h
@@ -26,7 +26,7 @@ constexpr double M_PI = 3.141592;
 #include "dlaf/common/assert.h"
 #include "dlaf/common/index2d.h"
 #include "dlaf/common/range2d.h"
-#include "dlaf/matrix.h"
+#include "dlaf/matrix/matrix.h"
 #include "dlaf/types.h"
 
 /// @file
