+++ test/include/dlaf_test/matrix/util_matrix_futures.h
@@ -18,7 +18,7 @@
 #include <hpx/local/future.hpp>
 
 #include "gtest/gtest.h"
-#include "dlaf/matrix.h"
+#include "dlaf/matrix/matrix.h"
 
 namespace dlaf {
 namespace matrix {
