+++ test/unit/auxiliary/mc/test_norm.cpp
@@ -17,7 +17,7 @@
 
 #include "dlaf/communication/communicator_grid.h"
 #include "dlaf/lapack/enum_output.h"
-#include "dlaf/matrix.h"
+#include "dlaf/matrix/matrix.h"
 #include "dlaf/util_matrix.h"
 
 #include "dlaf_test/comm_grids/grids_6_ranks.h"
