+++ test/unit/factorization/mc/test_cholesky.cpp
@@ -11,7 +11,7 @@
 
 #include "gtest/gtest.h"
 #include "dlaf/communication/communicator_grid.h"
-#include "dlaf/matrix.h"
+#include "dlaf/matrix/matrix.h"
 #include "dlaf_test/comm_grids/grids_6_ranks.h"
 #include "dlaf_test/matrix/util_matrix.h"
 #include "dlaf_test/util_types.h"
