+++ test/unit/matrix/test_util_matrix.cpp
@@ -17,7 +17,7 @@
 
 #include "dlaf/communication/communicator_grid.h"
 #include "dlaf/communication/sync/basic.h"
-#include "dlaf/matrix.h"
+#include "dlaf/matrix/matrix.h"
 
 #include "dlaf_test/comm_grids/grids_6_ranks.h"
 #include "dlaf_test/matrix/util_matrix.h"
