+++ test/unit/solver/mc/test_triangular.cpp
@@ -13,7 +13,7 @@
 #include <sstream>
 #include <tuple>
 #include "gtest/gtest.h"
-#include "dlaf/matrix.h"
+#include "dlaf/matrix/matrix.h"
 #include "dlaf_test/comm_grids/grids_6_ranks.h"
 #include "dlaf_test/matrix/util_matrix.h"
 #include "dlaf_test/matrix/util_matrix_blas.h"
